requirements.txt
fastapi
uvicorn[standard]
sqlmodel
jinja2
python-multipart
psycopg2-binary
python-dotenv
